refactor: moved fixtures file to __data__ dir
